requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
asyncpg==0.30.0
alembic==1.13.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1